Refactor get all files
